+++ src/extra/signal/__init__.py
@@ -1,0 +1,4 @@
+
+from ._ftd import cfd, dled, config_ftd_interpolation, \
+    EdgeInterpolation  # noqa
+from ._interpolation import sinc_interpolate  # noqa